refactor(bookmarks): split bookmark form modules
